Move the engine cleanup tests out of the LookupIndex suite
Five of the seven tests in test_lookup_index_context_manager.py exercise
engine.py's owns_index bookkeeping, not LookupIndex, so they belong under
test_transformer/. The two that genuinely cover the context manager stay.

Deliberately keeping test_lookup_index_edge_cases.py, which #298 proposed
deleting: its coverage rows really are byte-identical, but it is the only
place asserting close lifecycle (close clears tables, operations after close
raise, double close is safe) and lookup_row's LIMIT 1 behaviour on duplicate
keys. Those lines are executed by other tests with other data, which is why
coverage can't see the loss.

Refs #298
